move resource/chatadapter cogito/gchatsink; remove empty pkg resource
PCI-2587
