+++ cogito/gchatsink_test.go
@@ -1,4 +1,4 @@
-package resource
+package cogito
 
 import (
 	"context"
